fix : Restructure ViewModel structure in project
- Move viewmodel classes to the corresponding UI packages
- Ensure correctness of imports in project files
