+++ app/src/test/java/com/github/lookupgroup27/lookup/model/profile/ProfileViewModelTest.kt
@@ -1,5 +1,6 @@
 package com.github.lookupgroup27.lookup.model.profile
 
+import com.github.lookupgroup27.lookup.ui.profile.ProfileViewModel
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.firestore.CollectionReference
